refactor: app list use case to use hook
